requirements.txt
openpyxl==3.1.5
pdf2image==1.17.0
pillow==10.4.0
azure-ai-formrecognizer==3.3.3
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
aiofiles==24.1.0
numpy==1.26.4
pdfminer.six==20231228
